Rearranged repository structure for multiple NetBeans projects in one repo
