+++ Assignments0102/nbproject/project.properties
@@ -3,8 +3,8 @@ annotation.processing.enabled.in.editor=false
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
-application.title=IST411
-application.vendor=redjen
+application.title=Assignments0102
+application.vendor=IST411Group4
 auxiliary.org-netbeans-modules-editor-indent.CodeStyle.project.expand-tabs=true
 auxiliary.org-netbeans-modules-editor-indent.CodeStyle.project.indent-shift-width=4
 auxiliary.org-netbeans-modules-editor-indent.CodeStyle.project.spaces-per-tab=4
@@ -36,7 +36,7 @@ debug.test.classpath=\
 dist.archive.excludes=
 # This directory is removed when the project is cleaned:
 dist.dir=dist
-dist.jar=${dist.dir}/IST411.jar
+dist.jar=${dist.dir}/Assignments0102.jar
 dist.javadoc.dir=${dist.dir}/javadoc
 endorsed.classpath=
 excludes=
