+++ Assignments0102/nbproject/project.xml
@@ -3,7 +3,7 @@
     <type>org.netbeans.modules.java.j2seproject</type>
     <configuration>
         <data xmlns="http://www.netbeans.org/ns/j2se-project/3">
-            <name>IST411</name>
+            <name>Assignments0102</name>
             <source-roots>
                 <root id="src.dir"/>
             </source-roots>
