+++ Assignments0102/src/hw02cats/Tree.java
